+++ apis/r/R/Factory.R
@@ -28,10 +28,9 @@
 #'   if it already exists, simply open it for writing.
 #' }
 #' @param platform_config Optional platform configuration.
-#' @param tiledbsoma_ctx Optional SOMATileDBContext.
+#' @param tiledbsoma_ctx Optional (DEPRECATED) SOMATileDBContext.
 #' @param tiledb_timestamp Optional Datetime (POSIXct) for TileDB timestamp.
-#' @param soma_context A SOMA context as created by
-#' \code{\link{soma_context}()}.
+#' @param soma_context Optional TileDB SOMA context.
 #'
 #' @return A new \link[tiledbsoma:SOMADataFrame]{SOMA data frame} stored at
 #' \code{uri} opened for writing.
@@ -68,10 +67,12 @@ SOMADataFrameCreate <- function(
   soma_context = NULL
 ) {
   ingest_mode <- match.arg(ingest_mode)
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMADataFrameCreate(tiledbsoma_ctx)")
   sdf <- SOMADataFrame$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context,
     tiledb_timestamp = tiledb_timestamp
   )
   ingest_mode <- switch(
@@ -123,10 +124,12 @@ SOMADataFrameOpen <- function(
     uri,
     tiledb_timestamp %||% "now"
   ))
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMADataFrameOpen(tiledbsoma_ctx)")
   sdf <- SOMADataFrame$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context,
     tiledb_timestamp = tiledb_timestamp
   )
   sdf$open(mode)
@@ -172,13 +175,16 @@ SOMASparseNDArrayCreate <- function(
   ingest_mode = c("write", "resume"),
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
   ingest_mode <- match.arg(ingest_mode)
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMASparseNDArrayCreate(tiledbsoma_ctx)")
   snda <- SOMASparseNDArray$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context,
     tiledb_timestamp = tiledb_timestamp
   )
   ingest_mode <- switch(
@@ -214,13 +220,16 @@ SOMASparseNDArrayOpen <- function(
   mode = "READ",
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMASparseNDArrayOpen(tiledbsoma_ctx)")
   snda <- SOMASparseNDArray$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   snda$open(mode)
   return(snda)
@@ -260,17 +269,20 @@ SOMADenseNDArrayCreate <- function(
   shape,
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
   soma_debug(sprintf(
     "[SOMADenseNDArrayCreate] tstamp (%s)",
     tiledb_timestamp %||% "now"
   ))
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMADenseNDArrayCreate(tiledbsoma_ctx)")
   dnda <- SOMADenseNDArray$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   dnda$create(type, shape, platform_config = platform_config)
   return(dnda)
@@ -296,13 +308,16 @@ SOMADenseNDArrayOpen <- function(
   mode = "READ",
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMADenseNDArrayOpen(tiledbsoma_ctx)")
   dnda <- SOMADenseNDArray$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   dnda$open(mode)
   return(dnda)
@@ -340,14 +355,17 @@ SOMACollectionCreate <- function(
   ingest_mode = c("write", "resume"),
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
   ingest_mode <- match.arg(ingest_mode)
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMACollectionCreate(tiledbsoma_ctx)")
   coll <- SOMACollection$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   ingest_mode <- switch(
     EXPR = ingest_mode,
@@ -384,13 +402,16 @@ SOMACollectionOpen <- function(
   mode = "READ",
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMACollectionOpen(tiledbsoma_ctx)")
   coll <- SOMACollection$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   coll$open(mode)
   return(coll)
@@ -440,14 +461,17 @@ SOMAMeasurementCreate <- function(
   ingest_mode = c("write", "resume"),
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
   ingest_mode <- match.arg(ingest_mode)
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMAMeasurementCreate(tiledbsoma_ctx)")
   meas <- SOMAMeasurement$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   ingest_mode <- switch(
     EXPR = ingest_mode,
@@ -482,13 +506,16 @@ SOMAMeasurementOpen <- function(
   mode = "READ",
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMAMeasurementOpen(tiledbsoma_ctx)")
   meas <- SOMAMeasurement$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   meas$open(mode)
   return(meas)
@@ -538,14 +565,17 @@ SOMAExperimentCreate <- function(
   ingest_mode = c("write", "resume"),
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
   ingest_mode <- match.arg(ingest_mode)
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMAExperimentCreate(tiledbsoma_ctx)")
   exp <- SOMAExperiment$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   ingest_mode <- switch(
     EXPR = ingest_mode,
@@ -579,13 +609,16 @@ SOMAExperimentOpen <- function(
   mode = "READ",
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
-  tiledb_timestamp = NULL
+  tiledb_timestamp = NULL,
+  soma_context = NULL
 ) {
+  soma_context <- get_soma_context(soma_context, tiledbsoma_ctx, what="SOMAExperimentOpen(tiledbsoma_ctx)")
   exp <- SOMAExperiment$new(
     uri,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
-    tiledb_timestamp = tiledb_timestamp
+    tiledb_timestamp = tiledb_timestamp,
+    soma_context = soma_context
   )
   exp$open(mode)
   return(exp)