+++ apis/r/R/soma_array_reader.R
@@ -56,8 +56,8 @@ soma_array_reader <- function(
     "'batch_size' must be character" = is.character(batch_size),
     "'result_order' must be character" = is.character(result_order),
     "'loglevel' must be character" = is.character(loglevel),
-    "'soma_context' must be external pointer or NULL" = is.null(soma_context) ||
-      inherits(soma_context, "externalptr")
+    "'soma_context' must be SOMAContext or NULL" = is.null(soma_context) ||
+      inherits(soma_context, "SOMAContext")
   )
 
   if (!is.null(dim_points)) {
@@ -70,7 +70,7 @@ soma_array_reader <- function(
   }
 
   if (is.null(soma_context)) {
-    soma_context <- soma_context()
+    soma_context <- SOMAContext$new()
   } # package-level cached instance
   if (is.null(timestamprange)) {
     soma_debug(
@@ -85,7 +85,7 @@ soma_array_reader <- function(
   }
   soma_array_reader_impl(
     uri,
-    soma_context,
+    soma_context$handle,
     colnames,
     qc,
     dim_points,