+++ apis/r/R/write_soma.R
@@ -9,7 +9,8 @@
 #' @template param-dots-method
 #' @param platform_config Optional \link[tiledbsoma:PlatformConfig]{platform
 #' configuration}.
-#' @param tiledbsoma_ctx Optional \code{\link{SOMATileDBContext}}.
+#' @param tiledbsoma_ctx Optional (DEPRECATED) \code{\link{SOMATileDBContext}}.
+#' @param soma_context Optional \code{\link{SOMAContext}}.
 #'
 #' @return The URI to the resulting \code{\link{SOMAExperiment}} generated from
 #' the data contained in \code{x}.
@@ -30,7 +31,8 @@ write_soma <- function(
   uri,
   ...,
   platform_config = NULL,
-  tiledbsoma_ctx = NULL
+  tiledbsoma_ctx = NULL,
+  soma_context = NULL
 ) {
   UseMethod(generic = "write_soma", object = x)
 }
@@ -93,6 +95,7 @@ write_soma.character <- function(
   ingest_mode = "write",
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
+  soma_context = NULL,
   relative = TRUE
 ) {
   sdf <- write_soma(
@@ -105,6 +108,7 @@ write_soma.character <- function(
     ingest_mode = ingest_mode,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context,
     relative = relative
   )
   sdf$set_metadata(uns_hint("1d"))
@@ -166,6 +170,7 @@ write_soma.data.frame <- function(
   ingest_mode = "write",
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
+  soma_context = NULL,
   relative = TRUE
 ) {
   stopifnot(
@@ -269,7 +274,8 @@ write_soma.data.frame <- function(
     domain = domain,
     ingest_mode = ingest_mode,
     platform_config = platform_config,
-    tiledbsoma_ctx = tiledbsoma_ctx
+    tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context
   )
   # Write values
   if (ingest_mode %in% c("resume")) {
@@ -322,6 +328,7 @@ write_soma.IterableMatrix <- function(
   shape = NULL,
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
+  soma_context = NULL,
   relative = TRUE
 ) {
   stopifnot(
@@ -376,11 +383,12 @@ write_soma.IterableMatrix <- function(
     shape = shape,
     ingest_mode = ingest_mode,
     platform_config = platform_config,
-    tiledbsoma_ctx = tiledbsoma_ctx
+    tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context
   )
   # TODO: Add support for resume-mode
   if (!is.null(x)) {
-    stride <- .block_size(ncol(x), tiledbsoma_ctx = array$tiledbsoma_ctx)
+    stride <- .block_size(ncol(x), array$soma_context)
     strider <- CoordsStrider$new(start = 1L, end = nrow(x), stride = stride)
     while (strider$has_next()) {
       idx <- as.integer(strider$next_element())
@@ -460,6 +468,7 @@ write_soma.matrix <- function(
   shape = NULL,
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
+  soma_context = NULL,
   relative = TRUE
 ) {
   stopifnot(
@@ -492,6 +501,7 @@ write_soma.matrix <- function(
       shape = shape,
       platform_config = platform_config,
       tiledbsoma_ctx = tiledbsoma_ctx,
+      soma_context = soma_context,
       relative = relative
     ))
   }
@@ -537,7 +547,8 @@ write_soma.matrix <- function(
     type = type %||% arrow::infer_type(x),
     shape = shape %||% dim(x),
     platform_config = platform_config,
-    tiledbsoma_ctx = tiledbsoma_ctx
+    tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context
   )
   # Write values
   array$write(x)
@@ -625,6 +636,7 @@ write_soma.TsparseMatrix <- function(
   shape = NULL,
   platform_config = NULL,
   tiledbsoma_ctx = NULL,
+  soma_context = NULL,
   relative = TRUE
 ) {
   stopifnot(
@@ -677,6 +689,7 @@ write_soma.TsparseMatrix <- function(
     ingest_mode = ingest_mode,
     platform_config = platform_config,
     tiledbsoma_ctx = tiledbsoma_ctx,
+    soma_context = soma_context,
     tiledb_timestamp = Sys.time()
   )
   # Write values